Significant re-writing of ESI code to start making it work with the new keckcode setup

+++ keckcode/esiredux/biastrim.py
@@ -1,4 +1,5 @@
-import pyfits,numpy
+import numpy
+from astropy.io import fits as pyfits
 
 def make_bias(filelist):
     count = 0
@@ -33,7 +34,7 @@ def make_bias(filelist):
 def biastrim(data,bias,bpm):
     # The bias floats, so the overscan region is better than bias frames
     if data.size>1500*3000:
-        data = data.T
+        data = data.T.astype(float)
         b1 = numpy.median(data[2074:2150],0)
         b2 = numpy.median(data[2152:2230],0)
         data[24:1048] -= b1
@@ -45,7 +46,7 @@ def biastrim(data,bias,bpm):
         data[24:1048] *= off
         return data[24:2072]*bpm*1.29
     else:
-        data = data.T
+        data = data.T.astype(float)
         b1 = numpy.median(data[1037:1075],0)
         b2 = numpy.median(data[1076:1115],0)
         data[12:524] -= b1
